refactor: reorganize view paths
