+++ frontend/src/router/index.js
@@ -3,19 +3,19 @@ import { useUserStore } from "../store";
 
 import LoginView from "../views/LoginView.vue";
 import MainLayout from "../layouts/MainLayout.vue";
-import DashboardView from "../views/DashboardView.vue";
-import CustomerCrawlView from "../views/CustomerCrawlView.vue";
-import PermissionView from "../views/PermissionView.vue";
-import CustomerManageView from "../views/CustomerManageView.vue";
-import SettingsView from "../views/SettingsView.vue";
-import ContentGenerateView from "../views/ContentGenerateView.vue";
-import EmailMarketingView from "../views/EmailMarketingView.vue";
-import SocialMediaView from "../views/SocialMediaView.vue";
-import TaskScheduleView from "../views/TaskScheduleView.vue";
-import BehaviorTrackView from "../views/BehaviorTrackView.vue";
-import ReportsView from "../views/ReportsView.vue";
-import HelpCenterView from "../views/HelpCenterView.vue";
-import NotificationCenterView from "../views/NotificationCenterView.vue";
+import DashboardView from "../views/dashboard/DashboardView.vue";
+import CustomerCrawlView from "../views/customer/CustomerCrawlView.vue";
+import PermissionView from "../views/system/PermissionView.vue";
+import CustomerManageView from "../views/customer/CustomerManageView.vue";
+import SettingsView from "../views/system/SettingsView.vue";
+import ContentGenerateView from "../views/content/ContentGenerateView.vue";
+import EmailMarketingView from "../views/marketing/EmailMarketingView.vue";
+import SocialMediaView from "../views/marketing/SocialMediaView.vue";
+import TaskScheduleView from "../views/scheduler/TaskScheduleView.vue";
+import BehaviorTrackView from "../views/analytics/BehaviorTrackView.vue";
+import ReportsView from "../views/report/ReportsView.vue";
+import HelpCenterView from "../views/support/HelpCenterView.vue";
+import NotificationCenterView from "../views/notification/NotificationCenterView.vue";
 import CampaignCenterView from "../views/CampaignCenterView.vue";
 import MenuManagement from "../views/system/MenuManagement.vue";
 const routes = [
