+++ frontend/src/views/analytics/BehaviorTrackView.vue
