+++ frontend/src/views/content/ContentGenerateView.vue
