+++ frontend/src/views/dashboard/DashboardView.vue
