+++ frontend/src/views/marketing/EmailMarketingView.vue
