+++ frontend/src/views/system/PermissionView.vue
